Initial version of blend2json.
This tool aims at exporting (parts of) a .blend file in JSON.

This makes it easy to 'read' the actual content of a .blend, to diff it, etc.

This is still very WIP, right it's OK to get the general structure of the .blend
(all its blocks), but data itself is not yet exportable.

There are also options to filter (based on keys & values regex matching) what is
exported to JSON, but this is not yet fully working as expected.